backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
watchdog==3.0.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
bcrypt==4.0.1
flask-limiter==3.5.0
email-validator==2.1.0.post1
cryptography==41.0.7
pycryptodome==3.19.1
